+++ config/environments/development.php
@@ -23,9 +23,17 @@ Config::define( 'SCRIPT_DEBUG', true );
 Config::define( 'WP_DEBUG_DISPLAY', true );
 Config::define( 'WP_DISABLE_FATAL_ERROR_HANDLER', true );
 
+ini_set( 'display_errors', '1' );
+
 /**
  * --------------------------------------------------------------------------
  * Enable plugin and theme updates and installation from the admin.
  * --------------------------------------------------------------------------
  */
 Config::define( 'DISALLOW_FILE_MODS', false );
+
+
+/**
+ * @since Brocooly 0.14.2
+ */
+Config::define( 'DISALLOW_INDEXING', true );