+++ web/app/themes/brocooly/config/timber.php
@@ -30,22 +30,22 @@ return [
 		/**
 		 * Help to include correct production bundle accordingly to manifest file
 		 */
-		'asset'   => 'asset',
+		'asset'     => 'asset',
 
 		/**
 		 * Get theme mod without customizer prefix
 		 */
-		'mod'     => 'mod',
+		'mod'       => 'mod',
 
 		/**
 		 * Get theme mod
 		 */
-		'rawmod'  => 'get_theme_mod',
+		'rawmod'    => 'get_theme_mod',
 
 		/**
 		 * Form action handler
 		 */
-		'handler' => 'action',
+		'handler'   => 'action',
 
 		/**
 		 * Carbon fields meta helpers