+++ web/app/themes/brocooly/inc/helpers.php
@@ -25,14 +25,14 @@ if ( ! function_exists( 'carbonPostMeta' ) ) {
 	/**
 	 * Carbon Fields post meta
 	 *
-	 * @param string $key | key ti retrieve
+	 * @param string   $key | key ti retrieve.
 	 * @param int|null $id | post ID.
-	 * @return void
+	 * @return mixed
 	 */
 	function carbonPostMeta( string $key, $id = null ) {
 		if ( class_exists( 'Carbon_Fields\Carbon_Fields' ) ) {
 			$postId = $id ?? get_the_ID();
-			$meta = carbon_get_post_meta( $postId, $key );
+			$meta   = carbon_get_post_meta( $postId, $key );
 			return $meta;
 		}