+++ CHANGELOG.md
@@ -5,13 +5,14 @@
 * **[Fix]**: fixed broken console commands
 * **[Fix]**: removed plugin disabler
 * **[Feat]**: added support for script localization
+* **[Feat]**: added .distignore file
 * **[Feat]**: now scripts may NOT be concat by default
 * **[Feat]**: added support for custom manifest file name
 * **[Feat]**: added `middlewareOnly()` method
 * **[Refactor]**: generic favicon now has additional check
 * **[Refactor]**: refactored `.htaccess` file
 
-Release Date: September 5, 2021
+Release Date: September 9, 2021
 
 ## Release 0.14.1